Synchrophasor: Fixed array out of bounds exception in the BodyImage property of the Ieee1344 FrequencyValue class.
git-tfs-id: [https://tfs.codeplex.com/tfs/TFS07]$/openpdc/Synchrophasor/Current Version;C41477

+++ Source/Libraries/TVA.PhasorProtocols/Ieee1344/FrequencyValue.cs
@@ -357,12 +357,12 @@ namespace TVA.PhasorProtocols.Ieee1344
 
                     if (definition.FrequencyIsAvailable)
                     {
-                        EndianOrder.BigEndian.CopyBytes(UnscaledFrequency, buffer, startIndex);
+                        EndianOrder.BigEndian.CopyBytes((short)UnscaledFrequency, buffer, startIndex);
                         startIndex += 2;
                     }
 
                     if (definition.DfDtIsAvailable)
-                        EndianOrder.BigEndian.CopyBytes(UnscaledDfDt, buffer, startIndex);
+                        EndianOrder.BigEndian.CopyBytes((short)UnscaledDfDt, buffer, startIndex);
                 }
 
                 return buffer;
